refactor: #159 tabs structure
add clean folder for clean tabs, add base folder for base structrs/methods for tabs
