+++ internal/tui/tabs/clean/filters_tab.go
@@ -1,4 +1,4 @@
-package tabs
+package clean
 
 import (
 	"strings"
