+++ internal/tui/tabs/clean/help_tab.go
@@ -1,4 +1,4 @@
-package tabs
+package clean
 
 import (
 	"strings"
